refactor: folder refactoring
