+++ src/i18n.py
@@ -1,7 +1,7 @@
 """Localization module"""
 
 import gettext
-from lib.__version__ import PACKAGE
+from __version__ import PACKAGE
 
 gettext.bindtextdomain(PACKAGE, "locales")
 gettext.textdomain(PACKAGE)
