+++ src/template/loader/plaintextloader.py
@@ -1,6 +1,6 @@
 """Plain/text Template Loader"""
 
-from lib.template.loader.abstractloader import AbstractTemplateLoader, register_loader
+from template.loader.abstractloader import AbstractTemplateLoader, register_loader
 
 
 @register_loader
