+++ src/template/loader/yamlloader.py
@@ -1,7 +1,7 @@
 """YAML Template Loader"""
 
 import yaml
-from lib.template.loader.abstractloader import AbstractTemplateLoader, register_loader
+from template.loader.abstractloader import AbstractTemplateLoader, register_loader
 
 
 @register_loader
