+++ tests/test_csv_loader.py
@@ -1,7 +1,7 @@
 """Test CSV template loader"""
 import logging
 import unittest
-from lib.template.loader.csvloader import CsvTemplateLoader
+from src.template.loader.csvloader import CsvTemplateLoader
 
 
 class TestCSVLoader(unittest.TestCase):
