+++ tests/test_json_loader.py
@@ -1,7 +1,7 @@
 """Test JSON template loader"""
 import logging
 import unittest
-from lib.template.loader.jsonloader import JsonTemplateLoader
+from src.template.loader.jsonloader import JsonTemplateLoader
 
 
 class TestJsonLoader(unittest.TestCase):
