+++ tests/test_plaintext_loader.py
@@ -1,7 +1,7 @@
 """Test CSV template loader"""
 import logging
 import unittest
-from lib.template.loader.plaintextloader import PlainTextTemplateLoader
+from src.template.loader.plaintextloader import PlainTextTemplateLoader
 
 
 class TestPlainTextLoader(unittest.TestCase):
